+++ authority-portal-backend/authority-portal-quarkus/src/test/kotlin/de/sovity/authorityportal/web/tests/services/organization/OrganizationUpdateApiServiceTest.kt
@@ -14,6 +14,8 @@
 package de.sovity.authorityportal.web.tests.services.organization
 
 import de.sovity.authorityportal.api.UiResource
+import de.sovity.authorityportal.api.model.UpdateOrganizationDto
+import de.sovity.authorityportal.api.model.UpdateOwnOrganizationDto
 import de.sovity.authorityportal.api.model.organization.OnboardingOrganizationUpdateDto
 import de.sovity.authorityportal.api.model.organization.OrganizationLegalIdTypeDto
 import de.sovity.authorityportal.db.jooq.Tables
@@ -128,6 +130,193 @@ class OrganizationUpdateApiServiceTest {
 
         // act
         val result = uiResource.updateOnboardingOrganization(request)
+
+        // assert
+        assertThat(result).isNotNull
+        assertThat(result.id).isEqualTo(dummyDevOrganizationId(0))
+
+        val actual = dsl.selectFrom(Tables.ORGANIZATION).where(Tables.ORGANIZATION.ID.eq(dummyDevOrganizationId(0))).fetchOne()
+        assertThat(actual).isNotNull
+        assertThat(actual!!.registrationStatus).isEqualTo(OrganizationRegistrationStatus.ACTIVE)
+        assertThat(actual.name).isEqualTo("Max's Organization")
+        assertThat(actual.description).isEqualTo("This is a test organization")
+        assertThat(actual.url).isEqualTo("https://www.test-organization.com")
+        assertThat(actual.businessUnit).isEqualTo("IT")
+        assertThat(actual.industry).isEqualTo("Software")
+        assertThat(actual.address).isEqualTo("Test Street, 12345, Test City, Test Country")
+        assertThat(actual.billingAddress).isEqualTo("Test Street, 12345, Test City, Test Country")
+        assertThat(actual.legalIdType).isEqualTo(OrganizationLegalIdTypeDto.COMMERCE_REGISTER_INFO.toDb())
+        assertThat(actual.commerceRegisterNumber).isEqualTo("HRB123456")
+        assertThat(actual.taxId).isNull()
+        assertThat(actual.commerceRegisterLocation).isEqualTo("Test City")
+        assertThat(actual.mainContactName).isEqualTo("Test Contact")
+        assertThat(actual.mainContactEmail).isEqualTo("contact@example.com")
+        assertThat(actual.mainContactPhone).isEqualTo("+1234567890")
+        assertThat(actual.techContactName).isEqualTo("Test Tech Contact")
+        assertThat(actual.techContactEmail).isEqualTo("techcontact@example.com")
+        assertThat(actual.techContactPhone).isEqualTo("+1234567890")
+    }
+
+    @Test
+    @TestTransaction
+    fun `update own organization fails because user is not admin of that organization`() {
+        // arrange
+        useDevUser(0, 0, setOf(Roles.UserRoles.PARTICIPANT_USER))
+
+        ScenarioData().apply {
+            organization(0, 0)
+            user(0, 0)
+            scenarioInstaller.install(this)
+        }
+
+        val request = UpdateOwnOrganizationDto(
+            description = "This is a test organization",
+            url = "https://www.test-organization.com",
+            businessUnit = "IT",
+            industry = "Software",
+            address = "Test Street, 12345, Test City, Test Country",
+            billingAddress = "Test Street, 12345, Test City, Test Country",
+            mainContactName = "Test Contact",
+            mainContactEmail = "contact@example.com",
+            mainContactPhone = "+1234567890",
+            techContactName = "Test Tech Contact",
+            techContactEmail = "techcontact@example.com",
+            techContactPhone = "+1234567890"
+        )
+
+        // act & assert
+        assertThatThrownBy {
+            uiResource.updateOwnOrganizationDetails(request)
+        }.isInstanceOf(NotAuthorizedException::class.java)
+    }
+
+    @Test
+    @TestTransaction
+    fun `update organization fails because user is not authority admin`() {
+        // arrange
+        useDevUser(0, 0, setOf(Roles.UserRoles.AUTHORITY_USER))
+
+        ScenarioData().apply {
+            organization(0, 0)
+            user(0, 0)
+            scenarioInstaller.install(this)
+        }
+
+        val request = UpdateOrganizationDto(
+            name = "Max's Organization",
+            description = "This is a test organization",
+            url = "https://www.test-organization.com",
+            businessUnit = "IT",
+            industry = "Software",
+            address = "Test Street, 12345, Test City, Test Country",
+            billingAddress = "Test Street, 12345, Test City, Test Country",
+            legalIdType = OrganizationLegalIdTypeDto.COMMERCE_REGISTER_INFO,
+            legalIdNumber = "HRB123456",
+            commerceRegisterLocation = "Test City",
+            mainContactName = "Test Contact",
+            mainContactEmail = "contact@example.com",
+            mainContactPhone = "+1234567890",
+            techContactName = "Test Tech Contact",
+            techContactEmail = "techcontact@example.com",
+            techContactPhone = "+1234567890"
+        )
+
+        // act & assert
+        assertThatThrownBy {
+            uiResource.updateOrganizationDetails(dummyDevOrganizationId(0), request)
+        }.isInstanceOf(NotAuthorizedException::class.java)
+    }
+
+    @Test
+    @TestTransaction
+    fun `update own organization as participant admin appropriately`() {
+        // arrange
+        val now = OffsetDateTime.now()
+
+        useDevUser(0, 0, setOf(Roles.UserRoles.PARTICIPANT_ADMIN))
+        useMockNow(now)
+
+        ScenarioData().apply {
+            organization(0, 0)
+            user(0, 0)
+            scenarioInstaller.install(this)
+        }
+
+        val request = UpdateOwnOrganizationDto(
+            description = "This is a test organization",
+            url = "https://www.test-organization.com",
+            businessUnit = "IT",
+            industry = "Software",
+            address = "Test Street, 12345, Test City, Test Country",
+            billingAddress = "Test Street, 12345, Test City, Test Country",
+            mainContactName = "Test Contact",
+            mainContactEmail = "contact@example.com",
+            mainContactPhone = "+1234567890",
+            techContactName = "Test Tech Contact",
+            techContactEmail = "techcontact@example.com",
+            techContactPhone = "+1234567890"
+        )
+
+        // act
+        val result = uiResource.updateOwnOrganizationDetails(request)
+
+        // assert
+        assertThat(result).isNotNull
+        assertThat(result.id).isEqualTo(dummyDevOrganizationId(0))
+
+        val actual = dsl.selectFrom(Tables.ORGANIZATION).where(Tables.ORGANIZATION.ID.eq(dummyDevOrganizationId(0))).fetchOne()
+        assertThat(actual).isNotNull
+        assertThat(actual!!.registrationStatus).isEqualTo(OrganizationRegistrationStatus.ACTIVE)
+        assertThat(actual.description).isEqualTo("This is a test organization")
+        assertThat(actual.url).isEqualTo("https://www.test-organization.com")
+        assertThat(actual.businessUnit).isEqualTo("IT")
+        assertThat(actual.industry).isEqualTo("Software")
+        assertThat(actual.address).isEqualTo("Test Street, 12345, Test City, Test Country")
+        assertThat(actual.billingAddress).isEqualTo("Test Street, 12345, Test City, Test Country")
+        assertThat(actual.mainContactName).isEqualTo("Test Contact")
+        assertThat(actual.mainContactEmail).isEqualTo("contact@example.com")
+        assertThat(actual.mainContactPhone).isEqualTo("+1234567890")
+        assertThat(actual.techContactName).isEqualTo("Test Tech Contact")
+        assertThat(actual.techContactEmail).isEqualTo("techcontact@example.com")
+        assertThat(actual.techContactPhone).isEqualTo("+1234567890")
+    }
+
+    @Test
+    @TestTransaction
+    fun `update organization as authority admin appropriately`() {
+        // arrange
+        val now = OffsetDateTime.now()
+
+        useDevUser(0, 0, setOf(Roles.UserRoles.AUTHORITY_ADMIN))
+        useMockNow(now)
+
+        ScenarioData().apply {
+            organization(0, 0)
+            user(0, 0)
+            scenarioInstaller.install(this)
+        }
+
+        val request = UpdateOrganizationDto(
+            name = "Max's Organization",
+            description = "This is a test organization",
+            url = "https://www.test-organization.com",
+            businessUnit = "IT",
+            industry = "Software",
+            address = "Test Street, 12345, Test City, Test Country",
+            billingAddress = "Test Street, 12345, Test City, Test Country",
+            legalIdType = OrganizationLegalIdTypeDto.COMMERCE_REGISTER_INFO,
+            legalIdNumber = "HRB123456",
+            commerceRegisterLocation = "Test City",
+            mainContactName = "Test Contact",
+            mainContactEmail = "contact@example.com",
+            mainContactPhone = "+1234567890",
+            techContactName = "Test Tech Contact",
+            techContactEmail = "techcontact@example.com",
+            techContactPhone = "+1234567890"
+        )
+
+        // act
+        val result = uiResource.updateOrganizationDetails(dummyDevOrganizationId(0), request)
 
         // assert
         assertThat(result).isNotNull