+++ authority-portal-frontend/src/app/pages/authority-organization-list-page/authority-organization-list-page/authority-organization-list-page.component.ts
@@ -13,8 +13,9 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {MatDialog} from '@angular/material/dialog';
 import {Title} from '@angular/platform-browser';
-import {Subject} from 'rxjs';
-import {takeUntil} from 'rxjs/operators';
+import {ActivatedRoute, Params, Router} from '@angular/router';
+import {Observable, Subject, combineLatest, merge, withLatestFrom} from 'rxjs';
+import {filter, skip, take, takeUntil} from 'rxjs/operators';
 import {Store} from '@ngxs/store';
 import {
   OrganizationOverviewEntryDto,
@@ -61,7 +62,7 @@ export class AuthorityOrganizationListPageComponent
   showDetail: boolean = false;
   slideOverConfig!: SlideOverConfig;
   componentToRender = AuthorityOrganizationDetailPageComponent;
-  filter: OrganizationRegistrationStatusDto | null = null;
+  organizationFilter: OrganizationRegistrationStatusDto | null = null;
   headerConfig!: HeaderBarConfig;
   filterBarConfig!: FilterBarConfig;
 
@@ -75,6 +76,8 @@ export class AuthorityOrganizationListPageComponent
     private slideOverService: SlideOverService,
     private globalStateUtils: GlobalStateUtils,
     private titleService: Title,
+    private activatedRoute: ActivatedRoute,
+    private router: Router,
   ) {
     this.titleService.setTitle('Participant Management');
   }
@@ -85,6 +88,8 @@ export class AuthorityOrganizationListPageComponent
     this.refresh();
     this.startListeningToState();
     this.startRefreshingOnEnvChange();
+
+    this.handleRouteParam();
   }
 
   initializeHeaderBar() {
@@ -144,7 +149,90 @@ export class AuthorityOrganizationListPageComponent
     });
   }
 
+  private handleRouteParam() {
+    const paramsChange: Observable<Params> = this.activatedRoute.params;
+    const stateReady: Observable<AuthorityOrganizationListPageState> =
+      this.store
+        .select<AuthorityOrganizationListPageState>(
+          AuthorityOrganizationListPageStateImpl,
+        )
+        .pipe(filter((state) => state.organizations.state === 'ready'));
+
+    // see https://stackoverflow.com/a/39104441
+    merge(
+      combineLatest([paramsChange, stateReady]).pipe(take(1)),
+      paramsChange.pipe(skip(1), withLatestFrom(stateReady)),
+    )
+      .pipe(takeUntil(this.ngOnDestroy$))
+      .subscribe(([params, state]) => {
+        if (params.organizationId) {
+          const entry: OrganizationOverviewEntryDto | undefined =
+            state.organizations.data.find(
+              (org) => org.id === params.organizationId,
+            );
+
+          if (entry) {
+            if (state.showDetail) {
+              this.updateSlideOver(
+                state.organizations.data.findIndex(
+                  (org) => org.id === entry.id,
+                ),
+              );
+            } else {
+              this.openSlideOver(entry);
+            }
+          } else {
+            this.router.navigate(['/authority/organizations']);
+          }
+        }
+      });
+  }
+
   openDetailPage(organization: OrganizationOverviewEntryDto) {
+    this.router.navigate(['/authority/organizations', organization.id]);
+  }
+
+  handleFilter(event: FilterQuery) {
+    this.organizationFilter = event[
+      'status'
+    ] as OrganizationRegistrationStatusDto;
+  }
+
+  handleNavigation(direction: SlideOverAction, currentConnectorId: string) {
+    let totalOrganizations = this.state.organizations.data.length;
+    let currentIndex = this.state.organizations.data.findIndex(
+      (organization) => organization.id === currentConnectorId,
+    );
+    let nextIndex = sliderOverNavigation(
+      direction,
+      currentIndex,
+      totalOrganizations,
+    );
+    this.updateSlideOver(nextIndex);
+    this.router.navigate([
+      '/authority/organizations',
+      this.state.organizations.data[nextIndex].id,
+    ]);
+  }
+
+  closeDetailPage() {
+    this.router.navigate(['/authority/organizations']);
+  }
+
+  inviteOrganization() {
+    this.dialog.open(AuthorityInviteNewOrganizationComponent, {
+      width: window.innerWidth > 640 ? '60%' : '100%',
+    });
+  }
+
+  ngOnDestroy$ = new Subject();
+  ngOnDestroy(): void {
+    this.closeSlideOver();
+    this.ngOnDestroy$.next(null);
+    this.ngOnDestroy$.complete();
+  }
+
+  private openSlideOver(organization: OrganizationOverviewEntryDto) {
     this.slideOverConfig = {
       childComponentInput: {
         id: organization.id,
@@ -158,46 +246,19 @@ export class AuthorityOrganizationListPageComponent
     this.store.dispatch(ShowOrganizationDetail);
   }
 
-  handleFilter(event: FilterQuery) {
-    this.filter = event['status'] as OrganizationRegistrationStatusDto;
-  }
-
-  handleNavigation(direction: SlideOverAction, currentConnectorId: string) {
-    let totalOrganizations = this.state.organizations.data.length;
-    let currentIndex = this.state.organizations.data.findIndex(
-      (organization) => organization.id === currentConnectorId,
-    );
-    let nextIndex = sliderOverNavigation(
-      direction,
-      currentIndex,
-      totalOrganizations,
-    );
+  private updateSlideOver(index: number) {
     this.slideOverConfig = {
       ...this.slideOverConfig,
       childComponentInput: {
-        id: this.state.organizations.data[nextIndex].id,
+        id: this.state.organizations.data[index].id,
       },
-      label: this.state.organizations.data[nextIndex].name,
+      label: this.state.organizations.data[index].name,
     };
     this.slideOverService.setSlideOverConfig(this.slideOverConfig);
   }
 
-  closeDetailPage() {
+  private closeSlideOver() {
     this.store.dispatch(CloseOrganizationDetail);
     this.slideOverService.slideOverReset();
-  }
-
-  inviteOrganization() {
-    this.dialog.open(AuthorityInviteNewOrganizationComponent, {
-      width: window.innerWidth > 640 ? '60%' : '100%',
-    });
-  }
-
-  ngOnDestroy$ = new Subject();
-
-  ngOnDestroy(): void {
-    this.closeDetailPage();
-    this.ngOnDestroy$.next(null);
-    this.ngOnDestroy$.complete();
   }
 }